refactor(adapter): the native-plugin reports live with the contract they query
`plugin_drift.go` sat in `internal/cli` but contained no cobra, no printing and
no flags — four pure queries over `adapter.PluginIngester`: which agents already
install a plugin themselves (`NativePluginOwners`, which seeds `native_agents`
at import), which declared plugins an agent installs AND agentsync also projects
there (`DuplicatedNativePlugins`), which enabled native plugins the canonical
never declares (`UndeclaredNativePlugins`), and the declared-set helper both
comparisons must share (`DeclaredPlugins`).

They move to `internal/adapter/nativeplugins.go`, beside the `PluginIngester`
interface and the `NativePlugin` value they consume. No new package edge:
`internal/adapter` already depends on `source` and `untrusted` and on nothing
else these need.

`internal/marketplace` — the issue's suggestion — was the alternative, and is
rejected: marketplace models manifests and projects them, and would have to take
a new dependency on `internal/adapter` to ask an agent what it has installed,
adding an edge the layering diagram does not have. Every one of these functions
opens with `reg.Lookup(name).(adapter.PluginIngester)`; the contract is the
subject, not the marketplace.

The gate tests move with them; `TestUnexaminedPluginAgents` and
`TestLosslessTargetingCaveat` stay in `internal/cli`, where their subjects are,
with the minimal ingester stub they need.
